navs: remove unused code & move it WIP
