+++ src/components/navs/AppNav/AppNavActions.vue
@@ -2,19 +2,25 @@
   <div>
     <div v-if="account" class="flex items-center">
       <AppNavActivityBtn />
-      <AppNavClaimBtn v-if="liquidityMiningSupported" />
+      <AppNavClaimBtn />
+      <AppNavBeets />
       <AppNavAccountBtn />
     </div>
-    <BalBtn
-      v-else
-      color="white"
-      :size="upToLargeBreakpoint ? 'md' : 'sm'"
-      @click="toggleWalletSelectModal"
-    >
-      <WalletIcon class="mr-2" />
-      <span class="hidden lg:inline-block" v-text="$t('connectWallet')" />
-      <span class="lg:hidden" v-text="$t('connect')" />
-    </BalBtn>
+    <div v-else class="flex">
+      <div class="mr-2">
+        <AppNavBeets />
+      </div>
+      <BalBtn
+        color="white"
+        :size="upToLargeBreakpoint ? 'md' : 'sm'"
+        @click="toggleWalletSelectModal"
+        class="mr-2"
+      >
+        <WalletIcon class="mr-2" />
+        <span class="hidden lg:inline-block" v-text="$t('connectWallet')" />
+        <span class="lg:hidden" v-text="$t('connect')" />
+      </BalBtn>
+    </div>
   </div>
 </template>
 
@@ -27,15 +33,17 @@ import useFathom from '@/composables/useFathom';
 import useBreakpoints from '@/composables/useBreakpoints';
 import useNumbers from '@/composables/useNumbers';
 
-import AppNavAccountBtn from './AppNavAccountBtn.vue';
-import AppNavClaimBtn from './AppNavClaimBtn.vue';
 import useWeb3 from '@/services/web3/useWeb3';
-import AppNavActivityBtn from './AppNavActivityBtn/AppNavActivityBtn.vue';
+import AppNavActivityBtn from '@/components/navs/AppNav/AppNavActivityBtn/AppNavActivityBtn.vue';
+import AppNavAccountBtn from '@/components/navs/AppNav/AppNavAccountBtn.vue';
+import AppNavClaimBtn from '@/components/navs/AppNav/AppNavClaimBtn.vue';
+import AppNavBeets from '@/components/navs/AppNav/AppNavBeets.vue';
 
 export default defineComponent({
   name: 'AppNavActions',
 
   components: {
+    AppNavBeets,
     AppNavAccountBtn,
     AppNavClaimBtn,
     AppNavActivityBtn
